Move .forMainButton() to extension
