+++ Packages/App/Sources/AppUITV/Views/Profile/ActiveProfileView.swift
@@ -74,7 +74,6 @@ struct ActiveProfileView: View {
                 .buttonStyle(.borderless)
             }
             .padding(.horizontal, 100)
-//            .padding(.top, 50)
 
             Spacer()
         }
@@ -169,23 +168,6 @@ private extension ActiveProfileView {
                 )
         }
         .focused($focusedField, equals: .switchProfile)
-    }
-}
-
-// MARK: - Local modifiers
-
-private extension View {
-    func forMainButton(withColor color: Color, focused: Bool, disabled: Bool) -> some View {
-        padding(.vertical, 25)
-            .background(disabled ? .gray : color)
-            .cornerRadius(50)
-            .font(.title3)
-            .foregroundColor(disabled ? .white.opacity(0.6) : .white)
-            .overlay(
-                RoundedRectangle(cornerRadius: 50)
-                    .fill(.white.opacity(focused ? 0.3 : 0.0))
-            )
-            .scaleEffect(focused ? 1.05 : 1.0)
     }
 }
